Add `check_count!` and a thread-local feature
When the thread-local feature is enabled (which is infeasible on some
targets that do not support TLS) we can count the coverage hits without
the concern of false positives that is present when a single atomic
global is used.

This allows us to also expose an API that allows users to assert an
exact number of mark hits, useful when counting e.g. double drops.