Move static to `hit!` to allow duplicate `check!`
Ideologically when using this library you will want each of the `hit!`
macros to be unique but there could be a desire to duplicate the
`check!` calls across different tests.

By creating the counter static in the `hit!` macro we ensure both of
these are enforced (at least within a crate).