feat: introduce mark survey
When initially designing a test, especially when retrofitting cov-mark
into an existing codebase, it can be quite handy to have a dryrun mode
during which we simply log which marks are hit.

This may help discover tests that cover similar marks, as well as help
select what other marks are sensible to add to the test.